moved RGB converter to qtui
